requirements.txt
faker==28.4.1
mysql-connector-python==9.0.0
python-dotenv==1.0.1